% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leftfootfinding.R
\name{leftfootfinding}
\alias{leftfootfinding}
\title{Left foot values}
\usage{
leftfootfinding(x, n)
}
\arguments{
\item{x}{A data set in the form of document-term matrix}

\item{n}{A constant value to fix the left foot value}
}
\value{
Left foot values for the input data set x.
}
\description{
Left foot value for triangular or trapezoidal membership function
}
\examples{
x<-matrix(c(12,9,14,11,21,16,15,24,20,17,14,11),nrow=4)
leftfootfinding(x,5)
#[1] 10 5 10 10
}
