% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leftshoulderfinding.R
\name{leftshoulderfinding}
\alias{leftshoulderfinding}
\title{Left shoulder values}
\usage{
leftshoulderfinding(a, b)
}
\arguments{
\item{a}{A constant value for fixing the left shoulder}

\item{b}{Middle values for the data set x}
}
\value{
Left shoulder values for the input data set x.
}
\description{
Left shoulder value for trapezoidal membership function
}
\examples{
x<-matrix(c(12,9,14,11,21,16,15,24,20,17,14,11),nrow=4)
mid<-midvalue(x)
leftshoulderfinding(2.5,mid)
#[1] 14.0 10.5 12.0 15.0
}
