% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midvalue.R
\name{midvalue}
\alias{midvalue}
\title{Middle values}
\usage{
midvalue(x)
}
\arguments{
\item{x}{A data set in the form of document-term matrix}
}
\value{
Middle values for the input data set x.
}
\description{
Middle value for triangular or trapezoidal membership function
}
\examples{
x<-matrix(c(12,9,14,11,21,16,15,24,20,17,14,11),nrow=4)
midvalue(x)
#[1] 16.5 13.0 14.5 17.5
}
