% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mn.R
\name{mn}
\alias{mn}
\title{Mean values}
\usage{
mn(x)
}
\arguments{
\item{x}{A data set in the form of document-term matrix}
}
\value{
Mean values for individual row of the input data set X.
}
\description{
Mean values of the data set for gaussian membership function
}
\examples{
x<-matrix(c(12,9,14,11,21,16,15,24,20,17,14,11),nrow=4)
mn(x)
#[1] 17.66667 14.00000 14.33333 15.33333
}
