% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rightfootfinding.R
\name{rightfootfinding}
\alias{rightfootfinding}
\title{Right foot values}
\usage{
rightfootfinding(x, n)
}
\arguments{
\item{x}{A data set in the form of document-term matrix}

\item{n}{A constant value to fix the right foot value}
}
\value{
Right foot values for the input data set x.
}
\description{
Right foot value for triangular or trapezoidal membership function
}
\examples{
x<-matrix(c(12,9,14,11,21,16,15,24,20,17,14,11),nrow=4)
rightfootfinding(x,5)
#[1] 25 20 20 25
}
