% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rightshoulderfinding.R
\name{rightshoulderfinding}
\alias{rightshoulderfinding}
\title{Right shoulder values}
\usage{
rightshoulderfinding(b, c)
}
\arguments{
\item{b}{Middle values for the data set x}

\item{c}{A constant value for fixing the right shoulder}
}
\value{
Right shoulder values for the input data set x.
}
\description{
right shoulder value for trapezoidal membership function
}
\examples{
x<-matrix(c(12,9,14,11,21,16,15,24,20,17,14,11),nrow=4)
mid<-midvalue(x)
rightshoulderfinding(mid,2.5)
#[1] 19.0 15.5 17.0 20.0
}
