% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simC.R
\name{simC}
\alias{simC}
\title{IFS similarity measure simC}
\usage{
simC(ma, na, mb, nb, k)
}
\arguments{
\item{ma}{IFS membership values for the data set x computed using either triangular or trapezoidal or guassian membership function}

\item{na}{IFS non-membership values for the data set x computed using either Sugeno and Terano's  or Yager's non-membership function}

\item{mb}{IFS membership values for the data set y computed using either triangular or trapezoidal or guassian membership function}

\item{nb}{IFS non-membership values for the data set y computed using either Sugeno and Terano's  or Yager's non-membership function}

\item{k}{A constant value, considered as 1}
}
\value{
The IFS similarity values of data set y with data set x
}
\description{
IFS similarity measure values using simC computation technique with membership, and non-membership values of two objects or set of objects.
}
\examples{
x<-matrix(c(12,9,14,11,21,16,15,24,20,17,14,11),nrow=4)
y<-matrix(c(11,21,6),nrow=1)
a<-mn(x)
b<-std(x)
a1<-mn(y)
b1<-std(y)
lam<-0.5
ma<-memG(a,b,x)
na<-nonmemS(ma,lam)
mb<-memG(a1,b1,y)
nb<-nonmemS(mb,lam)
k<-1
simC(ma,na,mb,nb,k)
#[1] 0.7005061 0.7011282 0.8783314 0.8783314
}
\references{
S.-M. Chen. Measures of similarity between vague sets. Fuzzy sets and Systems, 74(2):217 - 223, 1995.
}
