% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simKKDKS.R
\name{simKKDKS}
\alias{simKKDKS}
\title{SFS similarity measure simKKDKS}
\usage{
simKKDKS(ma, na, mb, nb, ia, ib, k)
}
\arguments{
\item{ma}{SFS membership values for the data set x computed using either triangular or trapezoidal or guassian membership function}

\item{na}{SFS non-membership values for the data set x computed using either Sugeno and Terano's  or Yager's non-membership function}

\item{mb}{SFS membership values for the data set y computed using either triangular or trapezoidal or guassian membership function}

\item{nb}{SFS non-membership values for the data set y computed using either Sugeno and Terano's  or Yager's non-membership function}

\item{ia}{SFS indeterminacy membership values for the data set x}

\item{ib}{SFS indeterminacy membership values for the data set y}

\item{k}{A constant value, considered as 1}
}
\value{
The SFS similarity values of data set y with data set x
}
\description{
SFS similarity measure values using simKKDKS computation technique with membership,non-membership, and indeterminacy membership values of two objects or set of objects.
}
\examples{
x<-matrix(c(12,9,14,11,21,16,15,24,20,17,14,11),nrow=4)
y<-matrix(c(11,21,6),nrow=1)
a<-mn(x)
b<-std(x)
a1<-mn(y)
b1<-std(y)
lam<-0.5
ma<-memG(a,b,x)
na<-nonmemS(ma,lam)
ia<-imemSFS(ma,na)
mb<-memG(a1,b1,y)
nb<-nonmemS(mb,lam)
ib<-imemSFS(mb,nb)
k<-1
simKKDKS(ma,na,mb,nb,ia,ib,k)
#[1] 0.5726216 0.3223250 0.2791418 0.2791418
}
\references{
M. J. Khan, P. Kumam, W. Deebani,W. Kumam, and Z. Shah. Distance and similarity measures for spherical fuzzy sets and their applications in selecting mega projects. Mathematics, 8(4):519, 2020.
}
