% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simZHFLL3.R
\name{simZHFLL3}
\alias{simZHFLL3}
\title{PFS similarity measure simZHFLL3}
\usage{
simZHFLL3(ma, na, mb, nb, ha, hb, k)
}
\arguments{
\item{ma}{PFS membership values for the data set x computed using either triangular or trapezoidal or guassian membership function}

\item{na}{PFS non-membership values for the data set x computed using either Sugeno and Terano's  or Yager's non-membership function}

\item{mb}{PFS membership values for the data set y computed using either triangular or trapezoidal or guassian membership function}

\item{nb}{PFS non-membership values for the data set y computed using either Sugeno and Terano's  or Yager's non-membership function}

\item{ha}{PFS hesitancy values for the data set x}

\item{hb}{PFS hesitancy values for the data set y}

\item{k}{A constant value, considered as 1}
}
\value{
The PFS similarity values of data set y with data set x
}
\description{
PFS similarity measure values using simZHFLL3 computation technique with membership, and non-membership values of two objects or set of objects.
}
\examples{
x<-matrix(c(12,9,14,11,21,16,15,24,20,17,14,11),nrow=4)
y<-matrix(c(11,21,6),nrow=1)
a<-mn(x)
b<-std(x)
a1<-mn(y)
b1<-std(y)
lam<-0.5
ma<-memG(a,b,x)
na<-nonmemS(ma,lam)
ha<-hmemPFS(ma,na)
mb<-memG(a1,b1,y)
nb<-nonmemS(mb,lam)
hb<-hmemPFS(mb,nb)
k<-1
simZHFLL3(ma,na,mb,nb,ha,hb,k)
#[1] 0.4742565 0.4823949 0.7745995 0.7745995
}
\references{
Q. Zhang, J. Hu, J. Feng, A. Liu, and Y. Li. New similarity measures of pythagorean fuzzy sets and their applications. IEEE Access, 7:138192 - 138202, 2019.
}
