% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/std.R
\name{std}
\alias{std}
\title{Standard deviation values}
\usage{
std(x)
}
\arguments{
\item{x}{A data set in the form of document-term matrix}
}
\value{
Standard deviation values for individual row of the input data set X.
}
\description{
Standard deviation of the data set for gaussian membership function
}
\examples{
x<-matrix(c(12,9,14,11,21,16,15,24,20,17,14,11),nrow=4)
std(x)
#[1] 4.9328829 4.3588989 0.5773503 7.5055535
}
